LayoutTests/media/track/track-constants-expected.txt

Test TextTrack constants.

EXPECTED (HTMLTrackElement.NONE == '0') OK
EXPECTED (HTMLTrackElement.LOADING == '1') OK
EXPECTED (HTMLTrackElement.LOADED == '2') OK
EXPECTED (HTMLTrackElement.ERROR == '3') OK
END OF TEST
